dts: Break all dependencies of DTS on ext
If DTS does not depend on include paths determined in /ext then we can
move dts next to kconfig. Moving it next to kconfig means that DTS
CONFIG_ variables will be available at the same time as kconfig's
CONFIG_ variables which is the intended design and beneficial from a
usability perspective.

Signed-off-by: Sebastian Boe <sebastian.boe@nordicsemi.no>