mbedtls: Added support for using an externally built mbedtls
Kbuild supported CONFIG_MBEDTLS_LIBRARY and
CONFIG_MBEDTLS_INSTALL_PATH to allow users to link in an externally
built mbedtls. This was not ported over to CMake, causing build
failures when it was kconfig-enabled.

This patch implements the same support. This support has been tested
as well as MBEDTLS_LIBRARY was tested in CI (Not at all).

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>