cmake: qemu: Support build-time configuration of the QEMU command
We introduce qemu_wrapper.py to support the use-case of configuring
the QEMU command line when 'make/ninja' is invoked (as opposed to when
CMake is invoked). qemu_wrapper.py uses environment variables to
achieve this.

From qemu_wrapper.py's internal documentation:

"""QEMU wrapper script

This script allows us to modify the QEMU command line without
It does so by wrapping QEMU (STDIN/STDERR and signals are 'forwarded'
to QEMU) and constructing a QEMU command line based on environment
variables and the CMake constructed QEMU command line.

Specifically, the environment variable 'run_args' will be appended to
the QEMU command when "make run" is invoked. The same applies for the
make target 'debugserver' (but with the environment variable
'debugserver_args')."""

Also, strings specified with the environment variable
'<target>_remove_args' will be removed from the QEMU command.

The make variables QEMU_INSTANCE, QEMU_PTY, and QEMU_PIPE exist to
support a use-case that is now supported by '<target>_[remove_]args'. The make
variables are kept for compatibility reasons, but are marked as
deprecated in the documentation.

<target>_[remove_]args should deprecate make variables because:

<target>_[remove_]args is portable to CMake-supporting IDEs, Ninja and
other CMake-generated build systems.

<target>_[remove_]args may eventually become a familiar pattern for
Zephyr users and therefore be easier to discover than obscure
undocumented Make variables.

<target>_[remove_]args is conceptually simpler than the Make
variables, as the Make variables require build script logic to
understand their semantics. The users model for <target>_[remove_]args
should be easy to understand, the flags given in <target>_args are
appended to the underlying command, target_remove_args are removed
from the underlying command.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>