kconfig: bluetooth: Introduce ranges to the RX and TX priorities
Invalid configurations should be detected during configuration instead
of during compilation whenever possible.

This patch removes a BUILD_ASSERT on the TX prioritiy being lower than
the RX priority and introduces two disjoint sets of valid ranges for
TX and RX priority.

The TX prioritiy is required to be at most 7, and the RX priority is
required to be at least 8. This is more restrictive than the original
build assert but on the other hand is enforcable by the configuration
system.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>