cmake: introduce ZEPHYR_TARGET
This is a global variable whose value is the target to use when
setting "global" properties for the zephyr build. New code should set
and get properties target from now on instead of relying on global
properties. ("Modern" CMake encourages thinking of build state in
terms of properties set on targets.)

Signed-off-by: Marti Bolivar <marti.bolivar@nordicsemi.no>