[nrf noup] drivers: usb_dc_dw: fix reset of soft disconnect
The stack first calls usb_dc_attach(), and a bit later configures
the control endpoints and callbacks. The driver implementation of
usb_dc_attach() resets soft disconnect and the host tries to enumerate
the device, but the control endpoint callbacks are not set at this
point.

Reset soft disconnect when stack calls usb_dc_ep_enable() for
the control IN endpoint.

Signed-off-by: Johann Fischer <johann.fischer@nordicsemi.no>